// src/chains/moonbeam/types/generated/v1606.ts
export type CurrencyId = CurrencyId_SelfReserve | CurrencyId_ForeignAsset | CurrencyId_LocalAssetReserve

export interface CurrencyId_SelfReserve {
    __kind: 'SelfReserve'
}

export interface CurrencyId_ForeignAsset {
    __kind: 'ForeignAsset'
    value: bigint
}

export interface CurrencyId_LocalAssetReserve {
    __kind: 'LocalAssetReserve'
    value: bigint
}